+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Styles of `<ol>` elements in `Task` [#97]
 - Removed `wavy` text decoration style of links because looks buggy
 - Improved `Table` styles to be no required to add `<th>` to all columns.
+- `Modal`: The close button conflicts with the extra content of the header
 
 ## [0.4.0] - 2020-03-06
 ### Added